@@ -14,16 +14,18 @@
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>2.0.3.RELEASE</version>
+		<version>2.4.5</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
 
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-		<java.version>1.8</java.version>
-		<axon.version>3.3.2</axon.version>
-	</properties>
+		<java.version>11</java.version>
+		<axon.version>4.4.8</axon.version>
+        <awaitility.version>3.1.6</awaitility.version>
+        <testcontainers.version>1.15.3</testcontainers.version>
+    </properties>
 
 	<dependencies>
         <dependency>
@@ -57,10 +59,17 @@
         <dependency>
             <groupId>org.awaitility</groupId>
             <artifactId>awaitility</artifactId>
-            <version>3.0.0-rc1</version>
+            <version>${awaitility.version}</version>
             <scope>test</scope>
         </dependency>
-	</dependencies>
+
+        <dependency>
+            <groupId>org.testcontainers</groupId>
+            <artifactId>junit-jupiter</artifactId>
+            <version>${testcontainers.version}</version>
+            <scope>test</scope>
+        </dependency>
+    </dependencies>
 
 	<build>
 		<plugins>
