@@ -1,20 +1,23 @@
 package com.example.axontrackingsagatest.domain.model;
 
 import com.example.axontrackingsagatest.domain.api.CreateSampleAggregateCommand;
+import com.example.axontrackingsagatest.domain.api.ModifySampleAggregateCommand;
 import com.example.axontrackingsagatest.domain.api.SampleAggregateCreated;
-import lombok.AccessLevel;
+import com.example.axontrackingsagatest.domain.api.SampleAggregateModified;
 import lombok.NoArgsConstructor;
 import org.axonframework.commandhandling.CommandHandler;
-import org.axonframework.commandhandling.model.AggregateIdentifier;
 import org.axonframework.eventsourcing.EventSourcingHandler;
+import org.axonframework.modelling.command.AggregateIdentifier;
+import org.axonframework.serialization.Revision;
 import org.axonframework.spring.stereotype.Aggregate;
 
 import java.util.UUID;
 
-import static org.axonframework.commandhandling.model.AggregateLifecycle.apply;
+import static org.axonframework.modelling.command.AggregateLifecycle.apply;
 
-@Aggregate
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Revision("1.0")
+@Aggregate(snapshotTriggerDefinition = "sampleSnapshotTrigger", snapshotFilter = "sampleSnapshotFilter")
+@NoArgsConstructor
 public class SampleAggregate {
     @AggregateIdentifier
     private UUID id;
@@ -22,6 +25,11 @@ public class SampleAggregate {
     @CommandHandler
     public SampleAggregate(CreateSampleAggregateCommand command) {
         apply(new SampleAggregateCreated(command.getId()));
+    }
+
+    @CommandHandler
+    public void handle(ModifySampleAggregateCommand command) {
+        apply(new SampleAggregateModified(command.getId()));
     }
 
     @EventSourcingHandler
